# api-docs.yml
...
swagger: '2.0'
info:
  description: 칵테일 공유 커뮤니티 마쉴랭의 API 명세서입니다.
  title: Maswillaeng
...
paths:
...
  "/api/duplicate-email":
    post:
      tags:
        - auth-controller
      summary: checkDuplicateEmail
      operationId: checkDuplicateEmailUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: email
          description: email
          required: true
          schema:
            type: object
            additionalProperties:
              type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/duplicate-nickname":
    post:
      tags:
        - auth-controller
      summary: checkDuplicateNickname
      operationId: checkDuplicateNicknameUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: nickName
          description: nickName
          required: true
          schema:
            type: object
            additionalProperties:
              type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  "/api/follower-list/{userId}":
    get:
      tags:
        - follow-controller
      summary: followerList
      operationId: followerListUsingGET
      produces:
        - "*/*"
      parameters:
        - name: userId
          in: path
          description: userId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/following-list/{userId}":
    get:
      tags:
        - follow-controller
      summary: followingList
      operationId: followingListUsingGET
      produces:
        - "*/*"
      parameters:
        - name: userId
          in: path
          description: userId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/following/{userId}":
    post:
      tags:
        - follow-controller
      summary: following
      operationId: followingUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - name: userId
          in: path
          description: userId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - follow-controller
      summary: followingDelete
      operationId: followingDeleteUsingDELETE
      produces:
        - "*/*"
      parameters:
        - name: userId
          in: path
          description: userId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/api/login":
    post:
      tags:
        - auth-controller
      summary: login
      operationId: loginUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: request
          description: request
          required: true
          schema:
            "$ref": "#/definitions/LoginRequestDto"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/logout":
    post:
      tags:
        - auth-controller
      summary: logout
      operationId: logoutUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/new-pid":
    get:
      tags:
        - follow-controller
      summary: followingPostList
      operationId: followingPostListUsingGET
      produces:
        - "*/*"
      parameters:
        - name: page
          in: query
          description: page
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  "/api/pwd-reset-mail":
    post:
      tags:
        - auth-controller
      summary: sendPwdResetMail
      operationId: sendPwdResetMailUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: email
          description: email
          required: true
          schema:
            type: object
            additionalProperties:
              type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  "/api/sign":
    post:
      tags:
        - auth-controller
      summary: sign
      operationId: signUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: userSignDto
          description: userSignDto
          required: true
          schema:
            "$ref": "#/definitions/UserJoinDto"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/update-token":
    get:
      tags:
        - auth-controller
      summary: updateAccessToken
      operationId: updateAccessTokenUsingGET
      produces:
        - "*/*"
      parameters:
        - name: refreshToken
          in: query
          description: refreshToken
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
definitions:
...
  LoginRequestDto:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    title: LoginRequestDto
...
  UserJoinDto:
    type: object
    properties:
      email:
        type: string
      nickName:
        type: string
      password:
        type: string
      userImage:
        type: string
    title: UserJoinDto
...